# src/Api/Swagger/View/openapi.yaml
openapi: 3.0.0
info:
  title: 'Websps API'
  version: '0.8'
paths:
...
  /api/employees/recommendations:
    post:
      tags:
        - Сотрудники
      summary: 'Создать рекомендацию сотруднику'
      description: '1с код рекомендатора берется из сессии, поэтому не передаем.'
      operationId: 978d0c741fbd45f2020f47fcc9ed2aa0
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddEmployeesRecommendationsDto'
      responses:
        '201':
          description: 'Успешно добавлено'
        '400':
          $ref:
            example: "\r\n{\r\n    \"error\": {\r\n        \"status\": 400,\r\n        \"user\": \"Скородумов Николай Валерьевич-УП00040092-12112\",\r\n        \"exception\": \"ValidationException\",\r\n        \"message\": \"Голосовать можно один раз в месяц. Вы уже голосовали.\",\r\n        \"file\": \"/var/www/html/www/src/src/Api/Additional/EmployeesRecommendations/EmployeesRecommendationsStorage.php\",\r\n        \"line\": 100,\r\n        \"previous\": null,\r\n        \"trace\": []\r\n    }\r\n}\r\n                    "
  '/api/employees/{recommender_code_1c}/recommendations':
    get:
      tags:
        - Сотрудники
      summary: 'Возвращает список все рекомендаций пользователя'
      description: ''
      operationId: 77072d9330ec6229f32f35070f2a5691
      parameters:
        -
          name: recommender_code_1c
          in: path
          required: true
          schema:
            schema: string
      responses:
        '200':
          content:
            application/json:
              schema: {  }
              example: "\r\n                    [\r\n  {\r\n    \"employee_name\": \"Ларионов Роман Сергеевич\",\r\n    \"employee_code_1c\": \"УП00035198\",\r\n    \"recommender_name\": \"Скородумов Николай Валерьевич\",\r\n    \"recommender_code_1c\": \"УП00040092\",\r\n    \"leadership_score\": \"5\",\r\n    \"skill_score\": \"5\",\r\n    \"comment\": \"Это пример комментария, который содержит более 30 символов и соответствует требованиям.\"\r\n  }\r\n]\r\n                    "
  /api/corporate-slider/categories:
    get:
      tags:
        - CorporateSlider
      summary: 'Возвращает список всех категорий с их слайдами'
      operationId: 2aa53be6b308b5337e1af3802f29f22a
      responses:
        '200':
          description: 'Список всех категорий и их слайдов'
          content:
            application/json:
              schema:
                properties:
                  categories: { type: array, items: { $ref: '#/components/schemas/CategoryResponseDto' } }
                type: object
    post:
      tags:
        - CorporateSlider
      summary: 'Создает новую категорию'
      operationId: 3ee7e5d58b34c38776d3c54008af9ca9
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CategoryCreateDto'
      responses:
        '201':
          description: 'Категория успешно создана'
          content:
            application/json:
              schema:
                properties:
                  category_uuid: { type: string }
                type: object
    delete:
      tags:
        - CorporateSlider
      summary: 'Удаляет категории по массивам UUID'
      operationId: 3766241ddf530cf587963a13c001d47c
      parameters:
        -
          name: category_uuids
          in: query
          description: 'Json строка с массивом uuids удаляемых категории'
          required: true
          example: '["uuid1", "uuid2"]'
      responses:
        '200':
          description: 'Категории успешно удалены'
  '/api/corporate-slider/categories/{category_uuid}':
    get:
      tags:
        - CorporateSlider
      summary: 'Получает категорию по её UUID'
      operationId: c48dc4e5207e2791b120d95995bf6c69
      parameters:
        -
          name: category_uuid
          in: path
          description: 'UUID категории'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  category: { $ref: '#/components/schemas/CategoryResponseDto' }
                type: object
    post:
      tags:
        - CorporateSlider
      summary: 'Обновляет категорию по uuid'
      operationId: c859358975f4f6f4b752f4585b231e24
      parameters:
        -
          name: category_uuid
          in: path
          description: 'UUID категории'
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CategoryUpdateDto'
      responses:
        '200':
          description: 'Категория успешно обновлена'
    delete:
      tags:
        - CorporateSlider
      summary: 'Удаляет категорию по её UUID'
      operationId: ef8defdbdb61decab5c50a38351aabd2
      parameters:
        -
          name: category_uuid
          in: path
          description: 'UUID удаляемой категории'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Категория успешно удалена'
...
components:
  schemas:
    AddEmployeesRecommendationsDto:
      description: 'DTO для создания рекомендации сотруднику'
      properties:
        employee_code_1c:
          required:
            - true
          type: string
          example: УП00035198
        leadership_score:
          required:
            - true
          type: integer
          maximum: 5
          minimum: 0
          example: 5
        skill_score:
          required:
            - true
          type: integer
          maximum: 5
          minimum: 0
          example: 5
        comment:
          required:
            - true
          type: string
          minLength: 30
          example: 'Это пример комментария, который содержит более 30 символов и соответствует требованиям.'
      type: object
    CategoryCreateDto:
      description: 'DTO категории'
      properties:
        category_name:
          description: 'Название категории'
          type: string
          example: 'Новая категория'
        slides:
          description: 'Слайды категории'
          type: array
          items:
            $ref: '#/components/schemas/SlideCreateDto'
          nullable: true
      type: object
    ImageCreateDto:
      description: 'DTO для изображения'
      properties:
        sort:
          description: 'Позиция изображения'
          type: integer
          example: 0
        file:
          description: 'Файл изображения'
          type: string
          format: binary
          nullable: true
      type: object
    SlideCreateDto:
      description: 'DTO для слайда'
      properties:
        sort:
          description: 'Позиция слайда'
          type: integer
          example: 0
        images:
          description: 'Изображения слайда. Либо 1, либо 4'
          type: array
          items:
            $ref: '#/components/schemas/ImageCreateDto'
      type: object
    CategoryResponseDto:
      description: 'DTO для ответа с категорией'
      properties:
        category_uuid:
          description: 'UUID категории'
          type: string
        name:
          description: 'Название категории'
          type: string
        slides:
          description: 'Слайды категории'
          type: array
          items:
            $ref: '#/components/schemas/SlideResponseDto'
          nullable: true
      type: object
    ImageResponseDto:
      description: 'DTO для ответа с изображением'
      properties:
        image_uuid:
          description: 'UUID изображения'
          type: string
        path:
          description: 'Путь к изображению'
          type: string
        name:
          description: 'Название изображения'
          type: string
        sort:
          description: 'Позиция изображения'
          type: integer
      type: object
    SlideResponseDto:
      description: 'DTO для ответа со слайдом'
      properties:
        slide_uuid:
          description: 'UUID слайда'
          type: string
        sort:
          description: 'Позиция слайда'
          type: integer
        images:
          description: 'Изображения слайда'
          type: array
          items:
            $ref: '#/components/schemas/ImageResponseDto'
          nullable: true
      type: object
    CategoryUpdateDto:
      description: 'DTO категории'
      properties:
        category_name:
          description: 'Название категории'
          type: string
          example: 'Новая категория'
        slide_uuids_to_remove:
          description: 'Json строка с массивом uuids слайдеров для удаления'
          type: string
          example: '["uuid1", "uuid2"]'
        slides:
          description: 'Слайды категории'
          type: array
          items:
            $ref: '#/components/schemas/SlideUpdateDto'
          nullable: true
      type: object
    ImageUpdateDto:
      description: 'DTO для изображения'
      properties:
        image_uuid:
          description: 'Uuid изображения'
          type: integer
          example: uuid
          nullable: true
        sort:
          description: 'Позиция изображения'
          type: integer
          example: 0
        file:
          description: 'Файл изображения'
          type: string
          format: binary
      type: object
    SlideUpdateDto:
      description: 'DTO для слайда'
      properties:
        slide_uuid:
          description: 'Uuid слайда'
          type: integer
          example: uuid
        sort:
          description: 'Позиция слайда'
          type: integer
          example: 0
        images:
          description: 'Изображения слайда. Либо 1, либо 4'
          type: array
          items:
            $ref: '#/components/schemas/ImageUpdateDto'
          nullable: true
      type: object
...